Move the src changes into the test
