+++ package.json
@@ -30,7 +30,6 @@
     "@google-cloud/paginator": "^6.0.0",
     "@google-cloud/precise-date": "^5.0.0",
     "apache-arrow": "^21.0.0",
-    "avsc": "^5.7.9",
     "core-js": "^3.41.0",
     "extend": "^3.0.2",
     "google-auth-library": "^10.0.0",
@@ -46,6 +45,7 @@
     "@types/node": "^22.13.14",
     "@types/sinon": "^17.0.4",
     "@types/uuid": "^10.0.0",
+    "avsc": "^5.7.9",
     "c8": "^10.1.3",
     "gapic-tools": "^1.0.1",
     "gts": "^6.0.2",
